refactor: restructure builder components
