+++ bathbot-model/src/command_fields.rs
@@ -1,65 +1,15 @@
-use std::{
-    fmt::{Debug, Formatter, Result as FmtResult},
-    str::FromStr,
-};
+use std::str::FromStr;
 
 use rosu_v2::prelude::{GameMode, Grade};
-use serde::{
-    de::{Deserializer, Error as DeError, Unexpected},
-    ser::Serializer,
-    Deserialize, Serialize,
-};
 use time::UtcOffset;
 use twilight_interactions::command::{CommandOption, CreateOption};
 
-use crate::deser::bool_as_u8;
-
-macro_rules! define_enum {
-    (
-        #[$enum_meta:meta]
-        pub enum $enum_name:ident {
-            $(
-                #[$variant_meta:meta]
-                $variant:ident = $discriminant:literal,
-            )*
-        }
-    ) => {
-        #[$enum_meta]
-        pub enum $enum_name {
-            $(
-                #[$variant_meta]
-                $variant = $discriminant,
-            )*
-        }
-
-        impl<'de> Deserialize<'de> for $enum_name {
-            fn deserialize<D: Deserializer<'de>>(d: D) -> Result<Self, D::Error> {
-                match u8::deserialize(d)? {
-                    $( $discriminant => Ok(Self::$variant), )*
-                    other => Err(DeError::invalid_value(
-                        Unexpected::Unsigned(u64::from(other)),
-                        &stringify!($enum_name),
-                    )),
-                }
-            }
-        }
-
-        impl Serialize for $enum_name {
-            fn serialize<S: Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
-                s.serialize_u8(*self as u8)
-            }
-        }
-    }
-}
-
-define_enum! {
-    #[derive(Copy, Clone, CommandOption, CreateOption, Debug, Eq, PartialEq)]
-    pub enum ShowHideOption {
-        #[option(name = "Show", value = "show")]
-        Show = 0,
-        #[option(name = "Hide", value = "hide")]
-        Hide = 1,
-    }
+#[derive(Copy, Clone, CommandOption, CreateOption, Debug, Eq, PartialEq)]
+pub enum ShowHideOption {
+    #[option(name = "Show", value = "show")]
+    Show = 0,
+    #[option(name = "Hide", value = "hide")]
+    Hide = 1,
 }
 
 #[derive(Copy, Clone, CommandOption, CreateOption, Eq, PartialEq)]
@@ -222,161 +172,4 @@ timezone_option! {
     P10, "UTC+10", 10, "10";
     P11, "UTC+11", 11, "11";
     P12, "UTC+12", 12, "12";
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-pub struct ScoreEmbedSettings {
-    pub image: ScoreEmbedImage,
-    pub pp: ScoreEmbedPp,
-    pub map_info: ScoreEmbedMapInfo,
-    pub footer: ScoreEmbedFooter,
-    pub buttons: ScoreEmbedButtons,
-    pub hitresults: ScoreEmbedHitResults,
-}
-
-impl Default for ScoreEmbedSettings {
-    fn default() -> Self {
-        Self {
-            image: ScoreEmbedImage::Thumbnail,
-            pp: ScoreEmbedPp::Max,
-            map_info: ScoreEmbedMapInfo {
-                len: true,
-                ar: true,
-                cs: true,
-                od: true,
-                hp: true,
-                bpm: true,
-                n_obj: false,
-                n_spin: false,
-            },
-            footer: ScoreEmbedFooter::WithMapRankedDate,
-            buttons: ScoreEmbedButtons {
-                pagination: true,
-                render: true,
-                miss_analyzer: true,
-            },
-            hitresults: ScoreEmbedHitResults::OnlyMisses,
-        }
-    }
-}
-
-define_enum! {
-    #[derive(Copy, Clone, CommandOption, CreateOption, Debug, Eq, PartialEq)]
-    pub enum ScoreEmbedHitResults {
-        #[option(name = "Full", value = "full")]
-        Full = 0,
-        #[option(name = "Only misses", value = "miss")]
-        OnlyMisses = 1,
-    }
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-pub struct ScoreEmbedButtons {
-    #[serde(with = "bool_as_u8")]
-    pub pagination: bool,
-    #[serde(with = "bool_as_u8")]
-    pub render: bool,
-    #[serde(with = "bool_as_u8")]
-    pub miss_analyzer: bool,
-}
-
-define_enum! {
-    #[derive(Copy, Clone, CommandOption, CreateOption, Debug, Eq, PartialEq)]
-    pub enum ScoreEmbedFooter {
-        #[option(name = "With score date", value = "score_date")]
-        WithScoreDate = 0,
-        #[option(name = "With score date", value = "ranked_date")]
-        WithMapRankedDate = 1,
-        #[option(name = "Hide", value = "hide")]
-        Hide = 2,
-    }
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-pub struct ScoreEmbedMapInfo {
-    #[serde(with = "bool_as_u8")]
-    pub len: bool,
-    #[serde(with = "bool_as_u8")]
-    pub ar: bool,
-    #[serde(with = "bool_as_u8")]
-    pub cs: bool,
-    #[serde(with = "bool_as_u8")]
-    pub od: bool,
-    #[serde(with = "bool_as_u8")]
-    pub hp: bool,
-    #[serde(with = "bool_as_u8")]
-    pub bpm: bool,
-    #[serde(with = "bool_as_u8")]
-    pub n_obj: bool,
-    #[serde(with = "bool_as_u8")]
-    pub n_spin: bool,
-}
-
-impl ScoreEmbedMapInfo {
-    pub fn show(self) -> bool {
-        let Self {
-            len,
-            ar,
-            cs,
-            od,
-            hp,
-            bpm,
-            n_obj,
-            n_spin,
-        } = self;
-
-        len || ar || cs || od || hp || bpm || n_obj || n_spin
-    }
-}
-
-define_enum! {
-    #[derive(Copy, Clone, CommandOption, CreateOption, Debug, Eq, PartialEq)]
-    pub enum ScoreEmbedImage {
-        #[option(name = "Image", value = "image")]
-        Image = 0,
-        #[option(name = "Thumbnail", value = "thumbnail")]
-        Thumbnail = 1,
-        #[option(name = "None", value = "none")]
-        None = 2,
-    }
-}
-
-impl ScoreEmbedImage {
-    pub fn from_value(value: &str) -> Option<Self> {
-        match value {
-            "image" => Some(Self::Image),
-            "thumbnail" => Some(Self::Thumbnail),
-            "none" => Some(Self::None),
-            _ => None,
-        }
-    }
-}
-
-define_enum! {
-    #[derive(Copy, Clone, CommandOption, CreateOption, Eq, PartialEq)]
-    pub enum ScoreEmbedPp {
-        #[option(name = "Max PP", value = "max_pp")]
-        Max = 0,
-        #[option(name = "If-FC PP", value = "if_fc")]
-        IfFc = 1,
-    }
-}
-
-impl ScoreEmbedPp {
-    pub fn from_value(value: &str) -> Option<Self> {
-        match value {
-            "max_pp" => Some(Self::Max),
-            "if_fc" => Some(Self::IfFc),
-            _ => None,
-        }
-    }
-}
-
-impl Debug for ScoreEmbedPp {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        match self {
-            Self::Max => f.write_str("Max PP"),
-            Self::IfFc => f.write_str("If-FC PP"),
-        }
-    }
 }
