+++ bathbot-model/src/lib.rs
@@ -16,6 +16,7 @@ mod twitch;
 mod user_stats;
 
 pub mod command_fields;
+pub mod embed_builder;
 pub mod rosu_v2;
 pub mod twilight_gateway;
 pub mod twilight_model;
