+++ bathbot/src/active/impls/single_score.rs
@@ -1,18 +1,28 @@
-use std::{fmt::Write, time::Duration};
+use std::{borrow::Cow, fmt::Write, time::Duration};
 
 use bathbot_model::{
-    command_fields::{ScoreEmbedImage, ScoreEmbedSettings},
+    embed_builder::{
+        EmoteTextValue, HitresultsValue, MapperValue, ScoreEmbedSettings, SettingValue,
+        SettingsImage, Value,
+    },
     rosu_v2::user::User,
 };
 use bathbot_psql::model::configs::ScoreData;
 use bathbot_util::{
     constants::{GENERAL_ISSUE, ORDR_ISSUE, OSU_BASE},
+    datetime::{HowLongAgoDynamic, HowLongAgoText, SecToMinSec, SHORT_NAIVE_DATETIME_FORMAT},
     fields,
     numbers::round,
     AuthorBuilder, CowUtils, EmbedBuilder, FooterBuilder, MessageBuilder,
 };
 use eyre::{Report, Result};
 use futures::future::BoxFuture;
+use rosu_pp::model::beatmap::BeatmapAttributes;
+use rosu_v2::{
+    model::{GameMode, Grade},
+    prelude::RankStatus,
+};
+use time::OffsetDateTime;
 use twilight_model::{
     channel::message::{
         component::{ActionRow, Button, ButtonStyle},
@@ -25,28 +35,28 @@ use twilight_model::{
     },
 };
 
-use super::embed_builder::Settings;
 use crate::{
     active::{
-        impls::{CachedRender, CachedRenderData},
+        impls::{embed_builder::ValueKind, CachedRender, CachedRenderData},
         pagination::{async_handle_pagination_component, handle_pagination_modal, Pages},
         ActiveMessages, BuildPage, ComponentResult, IActiveMessage,
     },
     commands::{
         osu::{OngoingRender, RenderStatus, RenderStatusInner, RENDERER_NAME},
-        utility::ScoreEmbedDataWrap,
+        utility::{ScoreEmbedData, ScoreEmbedDataWrap},
     },
     core::{buckets::BucketName, Context},
+    embeds::HitResultFormatter,
     manager::{redis::RedisData, OwnedReplayScore, ReplayScore},
     util::{
         interaction::{InteractionComponent, InteractionModal},
+        osu::{GradeFormatter, ScoreFormatter},
         Authored, Emote, MessageExt,
     },
 };
 
 pub struct SingleScorePagination {
     pub settings: ScoreEmbedSettings,
-    pub new_settings: Settings,
     scores: Box<[ScoreEmbedDataWrap]>,
     score_data: ScoreData,
     username: Box<str>,
@@ -70,7 +80,6 @@ impl SingleScorePagination {
 
         Self {
             settings,
-            new_settings: Default::default(),
             scores,
             score_data,
             username: Box::from(user.username()),
@@ -92,9 +101,7 @@ impl SingleScorePagination {
     ) -> Result<BuildPage> {
         let score = &*self.scores[self.pages.index()].get_mut().await?;
 
-        let (name, value, footer_text) = self.new_settings.apply(score, self.score_data, mark_idx);
-
-        let fields = fields![name, value, false];
+        let embed = apply_settings(&self.settings, score, self.score_data, mark_idx);
 
         let title = format!(
             "{} - {} [{}] [{}★]",
@@ -140,24 +147,11 @@ impl SingleScorePagination {
             data.append_to_description(&score.score, &score.map, &mut description);
         }
 
-        let mut builder = EmbedBuilder::new()
+        let builder = embed
             .author(self.author.clone())
             .description(description)
-            .fields(fields)
             .title(title)
             .url(url);
-
-        match self.settings.image {
-            ScoreEmbedImage::Image => builder = builder.image(score.map.cover()),
-            ScoreEmbedImage::Thumbnail => builder = builder.thumbnail(score.map.thumbnail()),
-            ScoreEmbedImage::None => {}
-        }
-
-        if let Some(footer_text) = footer_text {
-            let emote = Emote::from(score.score.mode).url();
-            let footer = FooterBuilder::new(footer_text).icon_url(emote);
-            builder = builder.footer(footer);
-        }
 
         Ok(BuildPage::new(builder, false).content(content))
     }
@@ -508,4 +502,620 @@ pub enum SingleScoreContent {
     SameForAll(String),
     OnlyForIndex { idx: usize, content: String },
     None,
+}
+
+fn apply_settings(
+    settings: &ScoreEmbedSettings,
+    data: &ScoreEmbedData,
+    score_data: ScoreData,
+    mark_idx: Option<usize>,
+) -> EmbedBuilder {
+    const SEP_NAME: &str = "\t";
+    const SEP_VALUE: &str = " • ";
+
+    let map_attrs = data.map.attributes().mods(data.score.mods.clone()).build();
+
+    let mut field_name = String::new();
+    let mut field_value = String::new();
+    let mut footer_text = String::new();
+
+    let mut writer = &mut field_name;
+
+    let first = settings.values.first().expect("at least one field");
+
+    let next = settings
+        .values
+        .get(1)
+        .filter(|next| next.y == 0)
+        .map(|value| &value.inner);
+
+    match (&first.inner, next) {
+        (
+            Value::Ar | Value::Cs | Value::Hp | Value::Od,
+            Some(Value::Ar | Value::Cs | Value::Hp | Value::Od),
+        ) => {
+            writer.push('`');
+
+            if mark_idx == Some(0) {
+                writer.push('*');
+            }
+
+            let _ = match first.inner {
+                Value::Ar => write!(writer, "AR: {}", round(map_attrs.ar as f32)),
+                Value::Cs => write!(writer, "CS: {}", round(map_attrs.cs as f32)),
+                Value::Hp => write!(writer, "HP: {}", round(map_attrs.hp as f32)),
+                Value::Od => write!(writer, "OD: {}", round(map_attrs.od as f32)),
+                _ => unreachable!(),
+            };
+
+            if mark_idx == Some(0) {
+                writer.push('*');
+            }
+        }
+        (Value::MapRankedDate, _) if data.map.ranked_date().is_none() => {}
+        _ => {
+            let mut value = Cow::Borrowed(first);
+
+            if matches!(&first.inner, Value::Mapper(mapper) if mapper.with_status)
+                && data.map.status() == RankStatus::Ranked
+                && data.map.ranked_date().is_some()
+                && settings
+                    .values
+                    .iter()
+                    .any(|value| ValueKind::from_setting(value) == ValueKind::MapRankedDate)
+            {
+                value = Cow::Owned(SettingValue {
+                    inner: Value::Mapper(MapperValue { with_status: false }),
+                    y: first.y,
+                });
+            }
+
+            if mark_idx == Some(0) {
+                writer.push_str("__");
+            }
+
+            write_field(&value, data, &map_attrs, score_data, writer);
+
+            if mark_idx == Some(0) {
+                writer.push_str("__");
+            }
+        }
+    }
+
+    for (window, i) in settings.values.windows(3).zip(1..) {
+        let [prev, curr, next] = window else {
+            unreachable!()
+        };
+
+        match (&prev.inner, &curr.inner, &next.inner) {
+            (Value::Grade, Value::Mods, _) if prev.y == curr.y => {
+                writer.push(' ');
+
+                if mark_idx == Some(i) {
+                    writer.push_str("__");
+                }
+
+                let _ = write!(writer, "+{}", data.score.mods);
+
+                if mark_idx == Some(i) {
+                    writer.push_str("__");
+                }
+            }
+            (
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+            ) if prev.y == curr.y && curr.y == next.y => {
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                let _ = match curr.inner {
+                    Value::Ar => write!(writer, "AR: {}", round(map_attrs.ar as f32)),
+                    Value::Cs => write!(writer, "CS: {}", round(map_attrs.cs as f32)),
+                    Value::Hp => write!(writer, "HP: {}", round(map_attrs.hp as f32)),
+                    Value::Od => write!(writer, "OD: {}", round(map_attrs.od as f32)),
+                    _ => unreachable!(),
+                };
+
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                writer.push(' ');
+            }
+            (
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                _,
+            ) if prev.y == curr.y => {
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                let _ = match curr.inner {
+                    Value::Ar => write!(writer, "AR: {}", round(map_attrs.ar as f32)),
+                    Value::Cs => write!(writer, "CS: {}", round(map_attrs.cs as f32)),
+                    Value::Hp => write!(writer, "HP: {}", round(map_attrs.hp as f32)),
+                    Value::Od => write!(writer, "OD: {}", round(map_attrs.od as f32)),
+                    _ => unreachable!(),
+                };
+
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                if curr.y < SettingValue::FOOTER_Y {
+                    writer.push('`');
+                }
+            }
+            (
+                _,
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                Value::Ar | Value::Cs | Value::Hp | Value::Od,
+            ) if curr.y == next.y => {
+                if prev.y == curr.y {
+                    let sep = if curr.y == 0 { SEP_NAME } else { SEP_VALUE };
+                    writer.push_str(sep);
+                } else if curr.y == SettingValue::FOOTER_Y {
+                    writer = &mut footer_text;
+                } else if prev.y == 0 {
+                    writer = &mut field_value;
+                } else {
+                    writer.push('\n');
+                }
+
+                if curr.y < SettingValue::FOOTER_Y {
+                    writer.push('`');
+                }
+
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                let _ = match curr.inner {
+                    Value::Ar => write!(writer, "AR: {}", round(map_attrs.ar as f32)),
+                    Value::Cs => write!(writer, "CS: {}", round(map_attrs.cs as f32)),
+                    Value::Hp => write!(writer, "HP: {}", round(map_attrs.hp as f32)),
+                    Value::Od => write!(writer, "OD: {}", round(map_attrs.od as f32)),
+                    _ => unreachable!(),
+                };
+
+                if mark_idx == Some(i) {
+                    writer.push('*');
+                }
+
+                writer.push(' ');
+            }
+            (_, Value::MapRankedDate, _) if data.map.ranked_date().is_none() => {}
+            _ => {
+                let mut value = Cow::Borrowed(curr);
+
+                if matches!(&curr.inner, Value::Mapper(mapper) if mapper.with_status)
+                    && data.map.status() == RankStatus::Ranked
+                    && data.map.ranked_date().is_some()
+                    && settings
+                        .values
+                        .iter()
+                        .any(|value| ValueKind::from_setting(value) == ValueKind::MapRankedDate)
+                {
+                    value = Cow::Owned(SettingValue {
+                        inner: Value::Mapper(MapperValue { with_status: false }),
+                        y: curr.y,
+                    });
+                }
+
+                if prev.y == curr.y {
+                    let sep = if curr.y == 0 { SEP_NAME } else { SEP_VALUE };
+                    writer.push_str(sep);
+                } else if curr.y == SettingValue::FOOTER_Y {
+                    writer = &mut footer_text;
+                } else if prev.y == 0 {
+                    writer = &mut field_value;
+                } else {
+                    writer.push('\n');
+                }
+
+                let mark = if value.y == SettingValue::FOOTER_Y {
+                    "*"
+                } else {
+                    "__"
+                };
+
+                if mark_idx == Some(i) {
+                    writer.push_str(mark);
+                }
+
+                write_field(&value, data, &map_attrs, score_data, writer);
+
+                if mark_idx == Some(i) {
+                    writer.push_str(mark);
+                }
+            }
+        }
+    }
+
+    let last_idx = settings.values.len() - 1;
+    let last = settings.values.get(last_idx).expect("at least one value");
+    let prev = last_idx
+        .checked_sub(1)
+        .and_then(|idx| settings.values.get(idx));
+
+    if !(ValueKind::from_setting(last) == ValueKind::MapRankedDate
+        && data.map.ranked_date().is_none())
+        && last_idx > 0
+    {
+        let mark = if last.y == SettingValue::FOOTER_Y {
+            "*"
+        } else {
+            "__"
+        };
+
+        if prev.is_some_and(|prev| prev.y != last.y) {
+            if last.y == SettingValue::FOOTER_Y {
+                writer = &mut footer_text;
+            } else if prev.is_some_and(|prev| prev.y == 0) {
+                writer = &mut field_value;
+            } else {
+                writer.push('\n');
+            }
+
+            let mut value = Cow::Borrowed(last);
+
+            if matches!(&last.inner, Value::Mapper(mapper) if mapper.with_status)
+                && data.map.status() == RankStatus::Ranked
+                && data.map.ranked_date().is_some()
+                && settings
+                    .values
+                    .iter()
+                    .any(|value| ValueKind::from_setting(value) == ValueKind::MapRankedDate)
+            {
+                value = Cow::Owned(SettingValue {
+                    inner: Value::Mapper(MapperValue { with_status: false }),
+                    y: last.y,
+                });
+            }
+
+            if mark_idx == Some(last_idx) {
+                writer.push_str(mark);
+            }
+
+            write_field(&value, data, &map_attrs, score_data, writer);
+
+            if mark_idx == Some(last_idx) {
+                writer.push_str(mark);
+            }
+        } else {
+            match (prev.map(|prev| &prev.inner), &last.inner) {
+                (Some(Value::Grade), Value::Mods) => {
+                    writer.push(' ');
+
+                    if mark_idx == Some(last_idx) {
+                        writer.push_str("__");
+                    }
+
+                    let _ = write!(writer, "+{}", data.score.mods);
+
+                    if mark_idx == Some(last_idx) {
+                        writer.push_str("__");
+                    }
+                }
+                (
+                    Some(Value::Ar | Value::Cs | Value::Hp | Value::Od),
+                    Value::Ar | Value::Cs | Value::Hp | Value::Od,
+                ) => {
+                    if mark_idx == Some(last_idx) {
+                        writer.push('*');
+                    }
+
+                    let _ = match last.inner {
+                        Value::Ar => write!(writer, "AR: {}", round(map_attrs.ar as f32)),
+                        Value::Cs => write!(writer, "CS: {}", round(map_attrs.cs as f32)),
+                        Value::Hp => write!(writer, "HP: {}", round(map_attrs.hp as f32)),
+                        Value::Od => write!(writer, "OD: {}", round(map_attrs.od as f32)),
+                        _ => unreachable!(),
+                    };
+
+                    if mark_idx == Some(last_idx) {
+                        writer.push('*');
+                    }
+
+                    writer.push('`');
+                }
+                _ => {
+                    let sep = if last.y == 0 { SEP_NAME } else { SEP_VALUE };
+                    writer.push_str(sep);
+
+                    let mut value = Cow::Borrowed(last);
+
+                    if matches!(&last.inner, Value::Mapper(mapper) if mapper.with_status)
+                        && data.map.status() == RankStatus::Ranked
+                        && data.map.ranked_date().is_some()
+                        && settings
+                            .values
+                            .iter()
+                            .any(|value| ValueKind::from_setting(value) == ValueKind::MapRankedDate)
+                    {
+                        value = Cow::Owned(SettingValue {
+                            inner: Value::Mapper(MapperValue { with_status: false }),
+                            y: last.y,
+                        });
+                    }
+
+                    if mark_idx == Some(last_idx) {
+                        writer.push_str(mark);
+                    }
+
+                    write_field(&value, data, &map_attrs, score_data, writer);
+
+                    if mark_idx == Some(last_idx) {
+                        writer.push_str(mark);
+                    }
+                }
+            }
+        }
+    }
+
+    let fields = fields![field_name, field_value, false];
+
+    let mut builder = EmbedBuilder::new().fields(fields);
+
+    match settings.image {
+        SettingsImage::Thumbnail => builder = builder.thumbnail(data.map.thumbnail()),
+        SettingsImage::Image => builder = builder.image(data.map.cover()),
+        SettingsImage::Hide => {}
+    }
+
+    if !footer_text.is_empty() {
+        let emote = Emote::from(data.score.mode).url();
+        let footer = FooterBuilder::new(footer_text).icon_url(emote);
+        builder = builder.footer(footer);
+    }
+
+    builder
+}
+
+const DAY: Duration = Duration::from_secs(60 * 60 * 24);
+
+fn write_field(
+    value: &SettingValue,
+    data: &ScoreEmbedData,
+    map_attrs: &BeatmapAttributes,
+    score_data: ScoreData,
+    writer: &mut String,
+) {
+    match &value.inner {
+        Value::Grade => {
+            let _ = if value.y == 0 {
+                write!(
+                    writer,
+                    "{}",
+                    GradeFormatter::new(data.score.grade, None, false),
+                )
+            } else if value.y == SettingValue::FOOTER_Y {
+                write!(writer, "{:?}", data.score.grade)
+            } else {
+                write!(
+                    writer,
+                    "{}",
+                    GradeFormatter::new(data.score.grade, data.score.legacy_id, true),
+                )
+            };
+
+            // The completion is very hard to calculate for `Catch` because
+            // `n_objects` is not correct due to juicestreams so we won't
+            // show it for that mode.
+            let is_fail = data.score.grade == Grade::F && data.score.mode != GameMode::Catch;
+
+            if is_fail {
+                let n_objects = data.map.n_objects();
+
+                let completion = if n_objects != 0 {
+                    100 * data.score.total_hits() / n_objects
+                } else {
+                    100
+                };
+
+                let _ = write!(writer, "@{completion}%");
+            }
+        }
+        Value::Mods => {
+            let _ = write!(writer, "+{}", data.score.mods);
+        }
+        Value::Score => {
+            let _ = write!(writer, "{}", ScoreFormatter::new(&data.score, score_data));
+        }
+        Value::Accuracy => {
+            let _ = write!(writer, "{}%", round(data.score.accuracy));
+        }
+        Value::ScoreDate => {
+            let score_date = data.score.ended_at;
+
+            if value.y == SettingValue::FOOTER_Y {
+                if OffsetDateTime::now_utc() < score_date + DAY {
+                    let _ = write!(writer, "{}", HowLongAgoText::new(&score_date));
+                } else {
+                    writer.push_str(&score_date.format(&SHORT_NAIVE_DATETIME_FORMAT).unwrap());
+                }
+            } else {
+                let _ = write!(writer, "{}", HowLongAgoDynamic::new(&score_date));
+            }
+        }
+        Value::Pp(pp) => {
+            let bold = if value.y < SettingValue::FOOTER_Y {
+                "**"
+            } else {
+                ""
+            };
+            let tilde = if value.y < SettingValue::FOOTER_Y {
+                "~~"
+            } else {
+                ""
+            };
+
+            let _ = write!(writer, "{bold}{:.2}", data.score.pp);
+
+            match (pp.max, data.if_fc_pp.filter(|_| pp.if_fc)) {
+                (true, Some(if_fc_pp)) => {
+                    let _ = write!(
+                        writer,
+                        "{bold}/{max:.2}PP {tilde}({if_fc_pp:.2}pp){tilde}",
+                        max = data.max_pp.max(data.score.pp)
+                    );
+                }
+                (true, None) => {
+                    let _ = write!(writer, "{bold}/{:.2}PP", data.max_pp.max(data.score.pp));
+                }
+                (false, Some(if_fc_pp)) => {
+                    let _ = write!(writer, "pp{bold} {tilde}({if_fc_pp:.2}pp){tilde}");
+                }
+                (false, None) => writer.push_str("pp**"),
+            }
+        }
+        Value::Combo(combo) => {
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push_str("**");
+            }
+
+            let _ = write!(writer, "{}x", data.score.max_combo);
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push_str("**");
+            }
+
+            if combo.max {
+                let _ = write!(writer, "/{}x", data.max_combo);
+            }
+        }
+        Value::Hitresults(hitresults) => {
+            let _ = match hitresults {
+                HitresultsValue::Full => write!(
+                    writer,
+                    "{}",
+                    HitResultFormatter::new(data.score.mode, data.score.statistics.clone())
+                ),
+                HitresultsValue::OnlyMisses if value.y < SettingValue::FOOTER_Y => {
+                    write!(
+                        writer,
+                        "{}{}",
+                        data.score.statistics.count_miss,
+                        Emote::Miss
+                    )
+                }
+                HitresultsValue::OnlyMisses => {
+                    write!(writer, "{} miss", data.score.statistics.count_miss)
+                }
+            };
+        }
+        Value::Length => {
+            let clock_rate = map_attrs.clock_rate as f32;
+            let seconds_drain = (data.map.seconds_drain() as f32 / clock_rate) as u32;
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push('`');
+            }
+
+            let _ = write!(writer, "{}", SecToMinSec::new(seconds_drain).pad_secs());
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push('`');
+            }
+        }
+        Value::Ar | Value::Cs | Value::Hp | Value::Od => {
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push('`');
+            }
+
+            let mut write = |name, value| write!(writer, "{name}: {}", round(value as f32));
+
+            let _ = match &value.inner {
+                Value::Ar => write("AR", map_attrs.ar),
+                Value::Cs => write("CS", map_attrs.cs),
+                Value::Hp => write("HP", map_attrs.hp),
+                Value::Od => write("OD", map_attrs.od),
+                _ => unreachable!(),
+            };
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push('`');
+            }
+        }
+        Value::Bpm(emote_text) => {
+            let clock_rate = map_attrs.clock_rate as f32;
+            let bpm = round(data.map.bpm() * clock_rate);
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push_str("**");
+            }
+
+            let _ = match emote_text {
+                EmoteTextValue::Emote if value.y < SettingValue::FOOTER_Y => {
+                    write!(writer, "{} {bpm}", Emote::Bpm)
+                }
+                EmoteTextValue::Text | EmoteTextValue::Emote => write!(writer, "{bpm} BPM"),
+            };
+
+            if value.y < SettingValue::FOOTER_Y {
+                writer.push_str("**");
+            }
+        }
+        Value::CountObjects(emote_text) => {
+            let n = data.map.n_objects();
+
+            let _ = match emote_text {
+                EmoteTextValue::Emote if value.y < SettingValue::FOOTER_Y => {
+                    write!(writer, "{} {n}", Emote::CountObjects)
+                }
+                EmoteTextValue::Text | EmoteTextValue::Emote => {
+                    write!(
+                        writer,
+                        "{n} object{plural}",
+                        plural = if n == 1 { "" } else { "s" }
+                    )
+                }
+            };
+        }
+        Value::CountSpinners(emote_text) => {
+            let n = data.map.n_spinners();
+
+            let _ = match emote_text {
+                EmoteTextValue::Emote if value.y < SettingValue::FOOTER_Y => {
+                    write!(writer, "{} {n}", Emote::CountSpinners)
+                }
+                EmoteTextValue::Text | EmoteTextValue::Emote => {
+                    write!(
+                        writer,
+                        "{n} spinner{plural}",
+                        plural = if n == 1 { "" } else { "s" }
+                    )
+                }
+            };
+        }
+        Value::MapRankedDate => {
+            if let Some(ranked_date) = data.map.ranked_date() {
+                writer.push_str("Ranked ");
+
+                if OffsetDateTime::now_utc() < ranked_date + DAY {
+                    let _ = if value.y == SettingValue::FOOTER_Y {
+                        write!(writer, "{}", HowLongAgoText::new(&ranked_date))
+                    } else {
+                        write!(writer, "{}", HowLongAgoDynamic::new(&ranked_date))
+                    };
+                } else if value.y == SettingValue::FOOTER_Y {
+                    writer.push_str(&ranked_date.format(&SHORT_NAIVE_DATETIME_FORMAT).unwrap());
+                } else {
+                    let _ = write!(writer, "<t:{}:f>", ranked_date.unix_timestamp());
+                }
+            }
+        }
+        Value::Mapper(mapper) => {
+            let creator = data.map.creator();
+
+            let _ = if mapper.with_status {
+                write!(writer, "{:?} mapset by {creator}", data.map.status())
+            } else {
+                write!(writer, "Mapset by {creator}")
+            };
+        }
+    }
 }
